add capture preview layer that stays in portrait for smooth transition between portrait and landscape ui orientation when starting a live stream
fixes #ios-3213

requires same named branch on iOS_Universal

LFLiveKit, when starting an AVCaptureSession, makes one one output connection to the session, which defaults to landscape.  It then has video configuration property outputImageOrientation that it transforms to using a GPU filter.  This approach doesn't work for our use case of starting the camera capture when ui is in portrait -- while LFLiveKit can transition from imageOutputOrientation portrait to landscape, it uses the GPU to do it, but so does the UI runtime to animate the UI orienation change, so the GPU output image is warped and frozen during the change.

To get around this problem, this fix adds a second output connection that only feeds a capture preview layer. More, the capture preview layer is configured with orientation portrait, and this doesn't change -- but that's what we want because during the UI transition from portrait to landscape and back again, we use the transition animation coordinatore to rotate the preview layer such that it doesn't even look like it moves.