Fixes ios-3435
The 3435 issue was that overlay content extended beneath the notch and off the bottom of the frame in the player on iPhoneX when in landscape.  That fix adjusted margins, but also changed the way the overlay view is generated that no longer requires the scaling, so this commit simply removes the scaling.

There is also in this commit some commented out debug code that was put in place to simulate a lost connection to Gozer so as to ensure correct handling when the connection is lost, or if Gozer itself closes the socket because it suffered an internal error, recovery of which required the streamer to initiate a new session.