Initial attempt at solving ios-3494
Implements first attempt at bandwidth adaptive video resolution changes: i.e. increases video resolution from 360p to 540p, or from 540p to 720p, or decreases in reverse order based on an assessment of streaming bandwidth.  Streaming bandwidth is measured by the adaptive video bitrate monitoring in LFStreamingBuffer.  The monitoring frequently samples the size of the queue of frames waiting to be sent; if queue size is consistently decreasing or unchanging, then video bitrate can safely be slightly increased; alternatively, if the queue size is consistently increasing, then bitrate must be decreased.  When bitrate reaches and consistently stays at the max video bitrate of the current output frame resolution, then video resolution is increased; alternatively if bitrate reaches and remains at the minimum bitrate of the current output resolution, then video resolution is lowered.

To accommodate this adaptive algorithm, significant changes were made to LFLiveKit which prior was coded for an unchanging video resolution, despite containing exposed properties in LFLiveVideoConfiguration that suggest innate support.  In actuality, the implementation of these properties’ setters did not propagate changes, much less maintain the consistency of the correlated configuration parameters (e.g. videoQuality and videoSize).

The most significant design change sets the AVCaptureSession to always capture 720p frames and then to render to a texture that is the target output size.  We chose this approach because changing avcapturesession on the fly resulted in a period of black frames.

For GPUImageVideoCamera this means that target output resolution must be tracked.  We create a new initializer that sets the capture session preset to 1280x720, and add new private property outputFrameSize_landscape that can be set with the new initializer, or is inferred in the others.  Then in LFVideoCapture we add new method setNewVideoConfiguration that sets the new property directly as well as updates the render sizes of the filters in the chain.

As mentioned, video bitrate is dynamically adjusted.  In LFStreamBuffer, we significantly reduce the interval at which the size of the frame-to-be-sent queue, down to .1 sec from 1 sec, and now every .5 sec (not 5 sec) we ascertain the trend in the queue size and recommend to the delegate whether to increase or decrease the bit rate.

LFLiveSession is the LFStreamBuffer delegate.  Changes to LFLiveSession include maintaining private properties for video configurations at low, med, and high, since they all have different resolutions and min/max bitrates.  We also set up 3 video encoders, one for each level.  When a new frame arrives to be encoded, it is sent to the appropriate encoder based on it’s frame height.  The LFStreamBuffer delegate callback is where the decision is made to just change the bitrate or to change resolutions.

LFLiveVideoConfiguration got reworked so that correlated parameters are computed rather than set without validation.

Note:
-because the filter chain source image is already 16x9, the crop filter is no longer needed, i.e. we can filter down directly to 360p

In LFStreamingBuffer:
-changed interval granularity to float in order to be able to sample to-be-sent frame queue size at sub-second values.  Now sample every .1 second and evaluate whether to modify video bitrate every .5 second.

in LFLiveVideoConfiguration:
-fixed issue when method updateConfigurationBasedOnVideoQuality: validates session preset and if necessary chooses closest supported resolution, code now preserves configuration consistency by appropriately adjusting other state properties video size, video bitrate, and video quality
-added video bitrate calculation based on industry standard min/max bits/pixel quality value
-adjusted access level of properties to readonly except viideoQuality because resolution, bitrate, frame rate and session preset must all be consistent relative to each other, but currently code setter methods enforced no such consistency
-removed unused properties videoMin/MaxFrameRate because we do not adjust frame rate

In LFLiveSessioin:
-expanded handling in socketBufferStatus:status: to increase/decrease video resolutions if bitrate has been at the maximum/minumum value for the current video configuration for 3 calls in a row

-in LFVideoCapture:
-added method to change configuration, which also resets filter chain output size

-in GPUImageVideoCamera:
-added public property outputFrameSize_landscape, as described above, and

-in LiveStreamHelper:
-set LFLiveSession instance property adaptiveBitrate to true to trigger adaptive bandwidth resolution
-added setting LFLiveSession instance property videoSizeRespectingAspectRatio: to false. The property is used to determine if instance owners see videoSize as one that fits within private videoSize in order to preserve aspect ratio, i.e. if _videoSize = 640,480 and videoSizeRespectingAspectRatio is set to true and sessionPreset has a 16x9 resolution, then property videoSize returns result of AVMakeRectWithAspectRatioInsideRect.  In our use case, tho, we set video configuration videoSize to 640x360, so setting AVMakeRectWithAspectRatioInsideRect is unnecessary; but LFLiveVideoConfiguration property videoSize is assignable, so setting is simply a safeguard